+++ tokenise.c
@@ -3,16 +3,16 @@
 
 #include "tokenise.h"
 
-int tokenise_string(char* input_string, char delim, char ***tokens)
+size_t tokenise_string(char* input_string, char delim, char ***tokens)
 {
     char *string = malloc(strlen(input_string) + 1);
     strcpy(string, input_string);
     size_t string_len = strlen(string);
     size_t previous_delim_location = 0;
 
-    int num_tokens = 0;
+    size_t num_tokens = 0;
 
-    int i;
+    size_t i;
     for (i = 0; i < string_len; i++)
     {
         if (string[i] == delim || string[i] == '\n')
